+++ src/controller/Attacker.java
@@ -1,4 +1,4 @@
-package model;
+package controller;
 
 /**
  *
