+++ src/controller/Target.java
@@ -1,4 +1,4 @@
-package model;
+package controller;
 /**
  * 
  */
